commplete operations for all models

+++ app/Http/Controllers/CommentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreCommentRequest;
 use App\Http\Requests\UpdateCommentRequest;
 use App\Interfaces\CommentRepositoryInterface;
+use App\Models\Comment;
 use Illuminate\Http\Request;
 
 class CommentController extends Controller
@@ -21,7 +22,9 @@ class CommentController extends Controller
 
     public function create(StoreCommentRequest $request)
     {
-        return $this->commentRepository->create($request->validated());
+        $user = auth('sanctum')->user();
+        $this->authorize('create', Comment::class);
+        return $this->commentRepository->create($request->validated(), $user);
     }
 
     public function findById(string $id)
@@ -31,11 +34,15 @@ class CommentController extends Controller
 
     public function update(string $id, UpdateCommentRequest $request)
     {
+        $comment = Comment::findOrFail($id);
+        $this->authorize('update', $comment);
         return $this->commentRepository->update($id, $request->validated());
     }
 
     public function delete(string $id)
     {
+        $comment = Comment::findOrFail($id);
+        $this->authorize('delete', $comment);
         return $this->commentRepository->delete($id);
     }
 }
