+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Interfaces\UserRepositoryInterface;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class UserController extends Controller
@@ -16,26 +17,34 @@ class UserController extends Controller
     }
     public function list()
     {
+        $this->authorize('list',User::class);
         return $this->userRepository->list();
     }
 
     public function create(StoreUserRequest $request)
     {
-        return $this->userRepository->create($request->validated());
+        $user = auth('sanctum')->user();
+        $this->authorize('create',User::class);
+        return $this->userRepository->create($request->validated(), $user);
     }
 
     public function findById(int $id)
     {
+        $this->authorize('findById',User::class);
         return $this->userRepository->findById($id);
     }
 
     public function update(int $id, UpdateUserRequest $request)
     {
+        $user = User::findOrFail($id);
+        $this->authorize('update', $user);
         return $this->userRepository->update($id, $request->validated());
     }
 
     public function delete(int $id)
     {
+        $user = User::findOrFail($id);
+        $this->authorize('delete', $user);
         return $this->userRepository->delete($id);
     }
 }
